18 November #1

+++ scripts/merge_json.sh
@@ -1,7 +1,9 @@
 #! /bin/sh
 
 output_file=$FILES/RunII/output/event_analysis/Data13TeV_SingleMuon2015BPromptReco/luminosities/total.json
-rm $output_file
+if [ -f $output_file ]; then
+    rm $output_file
+fi
 for file in $FILES/RunII/output/event_analysis/Data13TeV_SingleMuon2015BPromptReco/luminosities/*.json
 do
     if [ "$file" == "$output_file" ]; then
